requirements.txt
requests
cx_Oracle
python-dotenv==1.2.1
sqlalchemy==2.0.46
greenlet==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing-extensions==4.15.0
tzdata==2025.3
jellyfish==0.8.2
numpy==1.24.4
pandas
rapidfuzz==1.4.1
aiohttp==3.9.1
psycopg2-binary==2.9.9